refactor: adjust the data catalog and pipelines
Setup for combining weekly results into a single results file.

File structure and pipeline are completed changed from previous versions
