+++ CHANGELOG.md
@@ -1,33 +1,15 @@
 # Changelog
 
-## [0.3.6] - 2020-09-06
-
-### [Unreleased]
-- Modified the file structure, and data catalog identifiers for annual projections
-
--[Breaking] removed the possibility of there being a "default" pipeline
-
-### Changed
-
-
 ## [Version] - 2020-09-06
 
-### Added
+### [Unreleased]
 
-- Add item
 
 ### Changed
 
-- Change item
+[0.4.0] - 2020-09-06
 
-### Fixed
+### Changed
 
-- Fixed item
-
-### Removed
-
-- Remove item
-
-### Security
-
-- Security item
+- Refactored in preparation for weekly projections/results
+- !Breaking removed the possibility of there being a "default" pipeline
