+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.3.6-dev1
+current_version = 0.4.0
 commit = False
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(\-(?P<release>[a-z]+)(?P<build>\d+))?
@@ -24,6 +24,13 @@ include_trailing_comma = True
 force_grid_wrap = 0
 use_parentheses = True
 line_length = 88
+known_kedro = kedro
+known_first_party = phantasyfootballer
+sections = FUTURE, STDLIB, THIRDPARTY, KEDRO, FIRSTPARTY, LOCALFOLDER
+no_lines_before = FUTURE, STDLIB, THIRDPARTY, LOCALFOLDER
+
+[black]
+exclude = setup.cfg
 
 [mypy]
 files = src
