+++ src/phantasyfootballer/io/CSVRemoteDataSet.py
@@ -1,7 +1,7 @@
 import importlib
 from typing import Any, Callable, Dict, List, Union
-
 import pandas as pd
+
 from kedro.io import AbstractDataSet
 
 MODULE_SEPARATOR = "."
