+++ src/phantasyfootballer/pipelines/data_import/__init__.py
@@ -1,5 +1,5 @@
-from .pipeline import (
+from .pipeline import (  # create_weekly_projections_pipeline,
     create_annual_projections_pipeline,
     create_annual_results_pipeline,
-    create_pipeline,
+    create_weekly_results_pipeline,
 )
