--- src/tests/pipelines/data_engineering/test_pipeline.py
@@ -1,68 +1,0 @@
-# Copyright 2020 QuantumBlack Visual Analytics Limited
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-# OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND
-# NONINFRINGEMENT. IN NO EVENT WILL THE LICENSOR OR OTHER CONTRIBUTORS
-# BE LIABLE FOR ANY CLAIM, DAMAGES, OR OTHER LIABILITY, WHETHER IN AN
-# ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF, OR IN
-# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-#
-# The QuantumBlack Visual Analytics Limited ("QuantumBlack") name and logo
-# (either separately or in combination, "QuantumBlack Trademarks") are
-# trademarks of QuantumBlack. The License does not grant you any right or
-# license to the QuantumBlack Trademarks. You may not use the QuantumBlack
-# Trademarks or any confusingly similar mark as a trademark for your product,
-# or use the QuantumBlack Trademarks in any other manner that might cause
-# confusion in the marketplace, including but not limited to in advertising,
-# on websites, or on software.
-#
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-This is a boilerplate test file for pipeline 'data_engineering'
-generated using Kedro 0.16.4.
-Please add your pipeline tests here.
-
-Kedro recommends using `pytest` framework, more info about it can be found
-in the official documentation:
-https://docs.pytest.org/en/latest/getting-started.html
-"""
-from pathlib import Path
-
-import pytest
-from kedro.framework.context import load_context
-from phantasyfootballer.common import Stats
-from phantasyfootballer.pipelines.data_engineering.nodes import (
-    _craft_scoring_dict,
-    _fetch_scoring_schemes,
-)
-
-
-@pytest.fixture
-def catalog():
-    proj_path = Path.cwd()  # point back to the root of the project
-    context = load_context(proj_path)
-    catalog = context.catalog
-    return catalog
-
-
-def test_filter(catalog):
-    df_ppr = catalog.load("scoring.ppr")
-    assert len(df_ppr) > 10
-
-
-def test_craft_scoring_scheme():
-    ppr_dict = _craft_scoring_dict("full_ppr")
-    assert ppr_dict[Stats.RCV_REC] == 1
-
-
-def test_fetch_scoring_schemes():
-    result = _fetch_scoring_schemes()
-    assert "standard" in result
